+++ activerecord/test/models/book.rb
@@ -10,6 +10,8 @@ class Book < ActiveRecord::Base
   enum status: [:proposed, :written, :published]
   enum read_status: {unread: 0, reading: 2, read: 3}
   enum nullable_status: [:single, :married]
+  enum print_status: [:in_print, :out_of_print]
+  enum review_status: {good: 'A+', bad: 'D', not_reviewed: nil}
 
   def published!
     super